requirements.txt
pillow==11.2.1
pygame==2.6.1
PyOpenGL==3.1.9
PyOpenGL-accelerate==3.1.9
